suricata-dash/requirements.txt
aiohttp==3.8.4
rich==13.4.0
python-dotenv==1.0.1
